Face_Recognition&HeartRate_Detection/requirements.txt
opencv-python~=4.5.5.64
scipy~=1.5.4
torchvision~=0.11.3
matplotlib~=3.3.4
pandas~=1.1.5

numpy~=1.19.2
pillow~=8.3.1
pymysql~=1.0.2
schedule~=1.1.0
camera~=1.3.0